added tests for basic intervention activity functionality + permissions
